backend/pyproject.toml
[project]
name = "quinn-flaskproject"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.12,<4"
dependencies = [
    "flask (>=3.1.3,<4.0.0)",
    "flask-smorest (>=0.46.2,<0.47.0)",
    "flask-jwt-extended (>=4.7.1,<5.0.0)",
    "flask-sqlalchemy (>=3.1.1,<4.0.0)",
    "flask-migrate (>=4.1.0,<5.0.0)",
    "flask-caching (>=2.3.1,<3.0.0)",
    "flask-cors (>=6.0.2,<7.0.0)",
    "redis (>=7.3.0,<8.0.0)",
    "gunicorn (>=25.1.0,<26.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "imagekitio (>=5.2.0,<6.0.0)",
    "python-slugify[unidecode] (>=8.0.4,<9.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)"
]
